feat: render activities using premade components and move the page to its own folder
